requirements.txt
streamlit>=1.33.0
opencv-python-headless>=4.9.0.80
numpy>=1.24.0
pillow>=10.0.0
pandas>=2.1.0
openpyxl>=3.1.0
scipy>=1.11.0
pytesseract>=0.3.10
pdf2image>=1.17.0
torch>=2.1.0
transformers>=4.39.0
